feat: add api request, locations screen and about screen

+++ lib/components/home/card_favorite/current_value.dart
@@ -1,10 +1,10 @@
 import 'package:air_pollution_app/utils/extensions.dart';
 import 'package:flutter/material.dart';
 
-import '../../glass_container.dart';
+import '../../shared/glass_container.dart';
 
 class CurrentValue extends StatelessWidget {
-  final int value;
+  final double value;
 
   const CurrentValue({
     super.key,
